+++ src/components/ui/icons/Icon.stories.tsx
@@ -13,6 +13,7 @@ import {
   Chat,
   Check,
   CheckRound,
+  Clear,
   Clock,
   Cog,
   ContentShare,
@@ -74,6 +75,7 @@ export const _Chat = () => <Chat width={text('width', '2rem')} />;
 export const _Check = () => <Check width={text('width', '2rem')} />;
 export const _CheckRound = () => <CheckRound width={text('width', '2rem')} />;
 export const _Clock = () => <Clock width={text('width', '2rem')} />;
+export const _Clear = () => <Clear width={text('width', '2rem')} />;
 export const _Cog = () => <Cog width={text('width', '2rem')} />;
 export const _Crown = () => <Crown width={text('width', '2rem')} />;